feat: add components and add new song
