+++ radio-songs-app/src/App.css
@@ -18,66 +18,3 @@
   border-radius: 5px;
   box-shadow: 0 0 10px rgba(0, 0, 0, 0.1);
 }
-
-.song-request-form {
-  display: flex;
-  flex-direction: row;
-  align-items: center;
-  gap: 10px;
-  margin-bottom: 20px;
-}
-
-.song-request-form input,
-.song-request-form select {
-  padding: 5px;
-  border: 1px solid #ccc;
-  border-radius: 5px;
-  flex-grow: 1; /* Make the input fields take up the available space */
-}
-
-.song-request-form button {
-  background-color: #007bff;
-  color: white;
-  padding: 5px 10px;
-  border: none;
-  cursor: pointer;
-  transition: background-color 0.3s;
-  border-radius: 5px;
-}
-
-.song-request-form button:hover {
-  background-color: #0056b3;
-}
-
-.song-list {
-  border: 1px solid #ccc;
-  padding: 10px;
-}
-
-.song-item {
-  display: flex;
-  justify-content: space-between;
-  padding: 5px;
-  border-bottom: 1px solid #ccc;
-}
-
-.header {
-  text-align: center;
-  padding: 20px;
-  color: white;
-  border-radius: 5px;
-  font-family: "Arial", sans-serif;
-  margin-bottom: 20px;
-}
-
-.header h1 {
-  margin: 0;
-  padding: 0;
-  font-size: 2rem;
-}
-
-.header p {
-  margin: 5px 0;
-  padding: 0;
-  font-size: 1rem;
-}
