+++ radio-songs-app/src/components/Header.css
@@ -1,0 +1,20 @@
+.header {
+  text-align: center;
+  padding: 20px;
+  color: white;
+  border-radius: 5px;
+  font-family: "Arial", sans-serif;
+  margin-bottom: 20px;
+}
+
+.header h1 {
+  margin: 0;
+  padding: 0;
+  font-size: 2rem;
+}
+
+.header p {
+  margin: 5px 0;
+  padding: 0;
+  font-size: 1rem;
+}
