+++ radio-songs-app/src/components/SongItem.css
@@ -1,0 +1,6 @@
+.song-item {
+  display: flex;
+  justify-content: space-between;
+  padding: 5px;
+  border-bottom: 1px solid #ccc;
+}
