+++ radio-songs-app/src/components/SongList.css
@@ -1,0 +1,4 @@
+.song-list {
+  border: 1px solid #ccc;
+  padding: 10px;
+}
